requirements.txt
coverage==7.11.0
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
PyYAML==6.0.3
rich==14.2.0
ruff==0.14.0
